backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
openai==1.76.0
firebase-admin==6.8.0
python-dotenv==1.0.1
Pillow==10.2.0
whitenoise==6.6.0
gunicorn==22.0.0 